Clean up the script startup code
