+++ utils.py
@@ -1,0 +1,52 @@
+from __future__ import absolute_import
+import os
+from .Qt.QtWidgets import QApplication, QSplashScreen, QDialog, QMainWindow
+
+
+def getUiFile(fileVar, subFolder="ui", uiName=None):
+    uiFolder, filename = os.path.split(fileVar)
+    if uiName is None:
+        uiName = os.path.splitext(filename)[0]
+    if subFolder:
+        uiFile = os.path.join(uiFolder, subFolder, uiName + ".ui")
+    return uiFile
+
+
+def rootWindow():
+    """
+    Returns the currently active QT main window
+    Only works for QT UIs like Maya
+    """
+    # for MFC apps there should be no root window
+    window = None
+    if QApplication.instance():
+        inst = QApplication.instance()
+        window = inst.activeWindow()
+        # Ignore QSplashScreen s, they should never be considered the root window.
+        if isinstance(window, QSplashScreen):
+            return None
+        # If the application does not have focus try to find A top level widget
+        # that doesn t have a parent and is a QMainWindow or QDialog
+        if window is None:
+            windows = []
+            dialogs = []
+            for w in QApplication.instance().topLevelWidgets():
+                if w.parent() is None:
+                    if isinstance(w, QMainWindow):
+                        windows.append(w)
+                    elif isinstance(w, QDialog):
+                        dialogs.append(w)
+            if windows:
+                window = windows[0]
+            elif dialogs:
+                window = dialogs[0]
+        # grab the root window
+        if window:
+            while True:
+                parent = window.parent()
+                if not parent:
+                    break
+                if isinstance(parent, QSplashScreen):
+                    break
+                window = parent
+    return window
